refactor: Move APIs out of utils, delete chains file
