+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/hashdit/metamask-snap.git"
   },
   "source": {
-    "shasum": "cCxVB/YVb6f/KZcKsrve1nN0UxTaCGknaJsOlF5ISBg=",
+    "shasum": "R+1rzaMtJC0e1P50Z3KpscoYcJ9XIseAPfyAAIS1cg0=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
